+++ src/k_rank.h
@@ -45,7 +45,7 @@ struct gpRank_t
 	UINT8 totalPlayers;
 
 	UINT8 position;
-	UINT8 skin;
+	UINT16 skin;
 
 	UINT32 winPoints;
 	UINT32 totalPoints;
